+++ app.py
@@ -18,6 +18,7 @@ from init import db, ma, bcrypt, jwt
 from blueprints.cli_bp import cli_bp
 from blueprints.auth_bp import auth_bp
 from blueprints.books_bp import books_bp
+from blueprints.comments_bp import comments_bp
 
 app = Flask(__name__)
 
@@ -32,6 +33,7 @@ bcrypt.init_app(app)
 app.register_blueprint(cli_bp)
 app.register_blueprint(auth_bp)
 app.register_blueprint(books_bp)
+app.register_blueprint(comments_bp)
     
 
     
@@ -65,64 +67,6 @@ def add_review():
 
     return jsonify(message='Review added successfully'), 201
 
-
-
-@app.route("/add_comment", methods=["POST"])
-@jwt_required()
-def create_comment():
-    
-    comment_data = request.json
-
-    comment_content = comment_data.get('comment_content')
-    review_id = comment_data.get('review_id')
-
-    user_email = get_jwt_identity()
-    stmt = db.select(User).filter_by(email=user_email)
-    user = db.session.scalar(stmt)
-
-    new_comment = Comment(
-        comment_content=comment_content,
-        username=user.username,
-        user_id=user.id,
-        review_id=review_id,
-        date_created=date.today()
-    )
-
-    db.session.add(new_comment)
-    db.session.commit()
-    return jsonify({'message': 'Comment created successfully'})
-
-
-@app.route("/delete_comment/<comment_id>", methods=["DELETE"])
-@jwt_required()
-def delete_comment(comment_id):
-    comment = Comment.query.get(comment_id)
-
-    if comment:
-        admin_or_owner_required(comment.user.email)
-        db.session.delete(comment)
-        db.session.commit()
-
-        return jsonify({'message': 'Comment deleted successfully'})
-    else:
-        return jsonify({'message': 'Comment not found'})
-    
-@app.route("/edit_comment/<comment_id>", methods=["PUT"])
-@jwt_required()
-def edit_comment(comment_id):
-    comment = Comment.query.get(comment_id)
-
-    if comment:
-        admin_or_owner_required(comment.user.email)
-        data = request.get_json()
-        
-        comment.comment_content = data.get('comment_content')
-
-        db.session.commit()
-
-        return jsonify({'message': 'Comment updated successfully'})
-    else:
-        return jsonify({'message': 'Comment not found'})
 
 @app.route('/update_review', methods=['PUT'])
 @jwt_required()
